requirements.txt
absl-py==0.11.0
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
six==1.15.0
tb-nightly==2.4.0a20201110
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
tf-estimator-nightly==2.4.0.dev2020102301
tf-nightly==2.5.0.dev20201110
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1

matplotlib~=3.3.3